Fatture e Stato fatture Services

+++ src/main/java/bw_team7/gestione_azienda_energia/fatture/payloads/FatturaDTO.java
@@ -3,6 +3,7 @@ package bw_team7.gestione_azienda_energia.fatture.payloads;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -12,11 +13,11 @@ public record FatturaDTO(
 
         @NotNull(message = "Il numero della fattura è obbligatorio")
         @Positive(message = "Il numero della fattura deve essere positivo")
-        Long numero,
+        Integer numero,
 
         @NotNull(message = "L'importo è obbligatorio")
         @Positive(message = "L'importo deve essere maggiore di zero")
-        Long importo,
+        BigDecimal importo,
 
         @NotNull(message = "L'ID del cliente è obbligatorio")
         UUID clienteId,
